pagination

+++ server/db/migrations/20220913072352-create-product.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       rating: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC(10,2),
       },
       description: {
         type: Sequelize.TEXT
